# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Trainning Documentation",
        "description": "API documentation for the project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.training.test/api",
            "description": "API Server"
        }
    ],
    "paths": {
        "/students": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get a list of students",
                "description": "Listado de los registros de estudiantes",
                "operationId": "7c586e93df821205d2c5af51e9dfd6d2",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "post": {
                "tags": [
                    "Students"
                ],
                "summary": "Create a new student",
                "operationId": "9e62661a4847ad2a4c15519a535ad5aa",
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            }
        },
        "/students/{id}": {
            "get": {
                "tags": [
                    "Students"
                ],
                "summary": "Get a student by ID",
                "operationId": "8b4ba20f528dd1c1b78cbbb5950dfc15",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Student not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Students"
                ],
                "summary": "Update a student by ID",
                "operationId": "df614f464255b5d0abc841caecc0ba78",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid request"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Students"
                ],
                "summary": "Delete a student by ID",
                "operationId": "8b7dedcaa57ee3874f0220652271d1a7",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Student not found"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Students",
            "description": "Students"
        }
    ]
}